+++ tox.ini
@@ -5,7 +5,7 @@ skipsdist=True
 [testenv]
 usedevelop=True
 commands =
-    py.test --cov=docker tests/test.py tests/utils_test.py
+    py.test --cov=docker tests/unit/
 deps =
     -r{toxinidir}/test-requirements.txt
     -r{toxinidir}/requirements.txt
